rotate head when clicking it
