+++ Interactive-ervaring/Assets/Animations/Crocodile 1.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 812b7eff5819ee146afa49b0a7d75ae3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
