+++ Interactive-ervaring/Assets/Animations/Crocodile.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f27ebc4a2024bc4ba6545fe8fb761b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
