+++ Interactive-ervaring/Assets/Animations/mouth.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c88f36762205530448f71c2dd93af3ad
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
