+++ Interactive-ervaring/Assets/Scripts/Camera scripts/RecCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e07febc3b8430846ad1a992129e4a87
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
