+++ Interactive-ervaring/Assets/Scripts/Model scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8983c7f80fd9b764e863252927fc8dd2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
