+++ Interactive-ervaring/Assets/Scripts/Model scripts/InteractionCrocodile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90a5804788780bd4484d2edd52e43fec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
