+++ Interactive-ervaring/ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,9 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/ARPage.unity
+    guid: 8ef5ee724a5d8f2479bb687876966e9e
+  - enabled: 1
     path: Assets/Scenes/AccountPage.unity
     guid: c1f9ebf51c1e1504a9291f1eb74e0398
   - enabled: 0
